// src/app/shared/dynamic-grid/enums/operation.enum.ts
export enum OperationEnum {
    GT = 'gt',
    LT = 'lt',
    EQUALS = 'equals',
    NOTEQUALS = 'notEquals',
    BEFORE = 'before',
    AFTER = 'after',
    CONTAINS = 'contains',
    NOTCONTAINS = 'notContains',
    BEGINS = 'begins',
    ENDS = 'ends',
    ISNULL = 'isNull',
    ISNOTNULL = 'isNotNull',
    EMPTY = 'empty',
    ISNOTEMPTY = 'isNotEmpty'
}